server/requirements.txt
asgiref==3.8.1
django==5.1.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
sqlparse==0.5.3
uvicorn[standard]==0.27.1
uuid==1.30
django-filter==24.3

# Authentication & Security
PyJWT==2.10.1
django-trench==0.3.1
pyotp==2.9.0

# HTTP & Networking
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiohttp-retry==2.8.3
aiosignal==1.3.2
requests==2.32.3
urllib3==2.3.0
django-cors-headers==4.6.0

# MFA & 2FA Providers
smsapi-client==2.9.6
twilio==9.4.2
yubico-client==1.13.0

# Utils
attrs==24.3.0
certifi==2024.12.14
charset-normalizer==3.4.1
frozenlist==1.5.0
multidict==6.1.0
propcache==0.2.1
yarl==1.18.3
nanoid==2.0.0

# Google API
google-api-python-client==2.114.0
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0

# DB
psycopg==3.2.4
psycopg-binary==3.2.4
psycopg-pool==3.2.4
